# site/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Run402 API",
    "version": "2.0.0",
    "description": "Full-stack infrastructure for AI agents. Postgres, REST API, auth, file storage, static site hosting, serverless functions. Pay-per-tier via x402 USDC on Base — subscribe once, then use wallet signatures for all operations within your tier.",
    "contact": {
      "name": "Run402",
      "url": "https://run402.com"
    },
...
  },
  "servers": [
    {
      "url": "https://api.run402.com",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/projects/v1/admin/{id}/functions": {
      "post": {
        "tags": [
          "Functions"
        ],
        "summary": "Deploy a serverless function",
        "operationId": "deployFunction",
        "description": "Deploys a JavaScript/TypeScript function to AWS Lambda. Returns an invocation URL.",
        "security": [
          {
            "serviceKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "code"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Function name"
                  },
                  "code": {
                    "type": "string",
                    "description": "Function source code"
                  },
                  "config": {
                    "type": "object",
                    "properties": {
                      "timeout": {
                        "type": "integer",
                        "description": "Timeout in seconds"
                      },
                      "memory": {
                        "type": "integer",
                        "description": "Memory in MB"
                      }
                    }
                  },
                  "deps": {
                    "type": "object",
                    "description": "NPM dependencies (name → version)",
                    "additionalProperties": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Function deployed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "url": {
                      "type": "string",
                      "format": "uri"
                    },
                    "status": {
                      "type": "string"
                    },
                    "runtime": {
                      "type": "string"
                    },
                    "timeout": {
                      "type": "integer"
                    },
                    "memory": {
                      "type": "integer"
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "get": {
        "tags": [
          "Functions"
        ],
        "summary": "List project functions",
        "operationId": "listFunctions",
        "security": [
          {
            "serviceKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Function list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "functions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string"
                          },
                          "url": {
                            "type": "string",
                            "format": "uri"
                          },
                          "runtime": {
                            "type": "string"
                          },
                          "timeout": {
                            "type": "integer"
                          },
                          "memory": {
                            "type": "integer"
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updated_at": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/projects/v1/admin/{id}/functions/{name}": {
      "patch": {
        "tags": [
          "Functions"
        ],
        "summary": "Update function metadata (schedule, timeout, memory) without redeploying code",
        "operationId": "patchFunction",
        "security": [
          {
            "serviceKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "schedule": {
                    "type": "string",
                    "nullable": true,
                    "description": "Cron expression (5-field) or null to remove"
                  },
                  "config": {
                    "type": "object",
                    "properties": {
                      "timeout": {
                        "type": "integer",
                        "description": "Timeout in seconds"
                      },
                      "memory": {
                        "type": "integer",
                        "description": "Memory in MB"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated function metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": { "type": "string" },
                    "runtime": { "type": "string" },
                    "timeout": { "type": "integer" },
                    "memory": { "type": "integer" },
                    "schedule": { "type": "string", "nullable": true },
                    "updated_at": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Function not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Functions"
        ],
        "summary": "Delete a function",
        "operationId": "deleteFunction",
        "security": [
          {
            "serviceKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Function deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "deleted"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/projects/v1/admin/{id}/functions/{name}/trigger": {
      "post": {
        "tags": [
          "Functions"
        ],
        "summary": "Manually trigger a function",
        "description": "Invoke a function immediately, same as a cron tick. Returns the function's response status and body. Updates schedule_meta if the function has a schedule.",
        "operationId": "triggerFunction",
        "security": [
          {
            "serviceKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Function invocation result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "description": "HTTP status code from the function"
                    },
                    "body": {
                      "type": "string",
                      "description": "Response body from the function"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Function not found"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/projects/v1/admin/{id}/functions/{name}/logs": {
      "get": {
        "tags": [
          "Functions"
        ],
        "summary": "Get function execution logs",
        "operationId": "getFunctionLogs",
        "security": [
          {
            "serviceKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tail",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            },
            "description": "Number of log lines to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Function logs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "logs": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/projects/v1/admin/{id}/secrets/{key}": {
      "delete": {
        "tags": [
          "Functions"
        ],
        "summary": "Delete a function secret",
        "operationId": "deleteSecret",
        "security": [
          {
            "serviceKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Secret deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "deleted"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apikey": {
        "type": "apiKey",
        "in": "header",
        "name": "apikey",
        "description": "Project anon_key (JWT). Returned when you provision a project."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "User access_token from /auth/v1/token. Pass as Authorization: Bearer <token>."
      },
      "serviceKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Project service_key (JWT with service_role). Returned when you provision a project. Pass as Authorization: Bearer <service_key>."
      },
      "adminKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-key",
        "description": "Server admin key. Required for admin operations."
      },
      "x402": {
        "type": "apiKey",
        "in": "header",
        "name": "X-PAYMENT",
        "description": "x402 payment header. The client library (e.g. @anthropic-ai/sdk x402 facilitator) handles this automatically. Payment is in USDC on Base (mainnet eip155:8453 or testnet eip155:84532)."
      },
      "walletAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Run402-Wallet",
        "description": "EIP-4361 wallet authentication. Requires three headers: X-Run402-Wallet (wallet address), X-Run402-Signature (signed message), X-Run402-Timestamp (unix timestamp). The wallet must have an active tier subscription."
      }
    },
...
    "responses": {
...
      "Unauthorized": {
        "description": "Missing or invalid authentication. For wallet-auth endpoints, provide X-Run402-Wallet, X-Run402-Signature, and X-Run402-Timestamp headers. For project endpoints, provide apikey header (anon_key) or Authorization: Bearer (service_key/access_token).",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}